Cleanup, add test for bbox
