+++ src/gldas/grid.py
@@ -83,3 +83,28 @@ def GLDAS025LandGrid():
 
 if __name__ == '__main__':
     GLDAS025LandGrid()
+
+
+def load_grid(land_points=True, bbox=None):
+    """
+    Load gldas grid.
+
+    Parameters
+    ----------
+    land_points : bool, optional (default: True)
+        Reshuffle only land points
+    bbox : tuple, optional (default: True)
+        (min_lat, min_lon, max_lat, max_lon)
+        Bounding box to limit reshuffling to.
+    """
+    if land_points:
+        subgrid = GLDAS025LandGrid()
+        if bbox is not None:
+            subgrid = subgrid4bbox(subgrid, *bbox)
+    else:
+        if bbox is not None:
+            subgrid = subgrid4bbox(GLDAS025Cellgrid(), *bbox)
+        else:
+            subgrid = None
+
+    return subgrid
