+++ tests/test_reshuffle.py
@@ -33,13 +33,10 @@ def test_reshuffle(landpoints, bbox, n_files_should):
                  ioclass_kws={'read_bulk': True, 'read_dates': False},
                  parameters=['SoilMoi0_10cm_inst', 'SoilMoi10_40cm_inst'])
 
-    ts = ds.read_ts(45.08, 15.1)
+    ts = ds.read(45.08, 15.1)
     ts_SM0_10_values_should = np.array([9.595, 9.593, 9.578,
                                         9.562, 9.555, 9.555, 9.556], dtype=np.float32)
     nptest.assert_allclose(ts['SoilMoi0_10cm_inst'].values, ts_SM0_10_values_should, rtol=1e-5)
     ts_SM10_40_values_should = np.array([50.065, 50.064, 50.062,
                                     50.060, 50.059, 50.059, 50.059], dtype=np.float32)
     nptest.assert_allclose(ts['SoilMoi10_40cm_inst'].values, ts_SM10_40_values_should,rtol=1e-5)
-
-if __name__ == '__main__':
-    test_reshuffle(False, True, 5)
